Add OS querying module
The os.sh script provides information related to the host OS. Simply
sourcing this script file leads to the following OS-related identifiers
to be populated which may be used by client code:
  - `OS_KERNEL`
  - `OS_DISTRO`
  - `OS_DISTRO_STR` (pretty version of the above)
  - `OS_VERSION`
  - `OS_VERSION_STR` (pretty version of the above)

Currently, this script only supports the following OS distributions.
  - Alpine Linux
  - Arch Linux
  - FreeBSD
  - Debian Ubunutu

There are plans to support more OS distributions in future. This script
relies on the msg.sh script being sourced, and is currently untested.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>