Upgrade message module to log module
The message module was a nice starting point, but it is also quite
helpful to be able to log the messages on to file, especially for
debugging issues on remote servers.

Accordingly, the message functions now all support logging if a log
destination is defined. In order to highlight that these messages are
being logged, the functions have also been renamed.

We have also introduced the `log_raw()` function, which is particularly
useful for logging the output of commonds.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>